requirements.txt
Django==6.0.2                  
djangorestframework==3.15.2
psycopg2-binary==2.9.11
djangorestframework-simplejwt==5.5.1  
redis==7.1.1                   
django-redis==6.0.0           
python-dotenv==1.2.1